requirements.txt
livekit-agents>=1.3.10
livekit-plugins-openai>=0.8.5
livekit-plugins-deepgram>=0.8.1
livekit-plugins-cartesia>=0.2.3
livekit-plugins-silero>=0.8.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
structlog>=24.1.0
livekit-plugins-sarvam
livekit-plugins-elevenlabs
